code/Automatic-Signature-Verification/Automatic-Signature-Verification/requirements.txt
appnope==0.1.3
backcall==0.2.0
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
matplotlib==3.3.4
nest-asyncio==1.5.6
numpy==1.19.5
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.7
python-dateutil==2.8.2
pyzmq==25.0.2
six==1.16.0
tornado==6.1
traitlets==4.3.3
wcwidth==0.2.6
